+++ iOS/Interfaces/SQLiteIOS.cs
@@ -40,14 +40,24 @@ namespace Dustbuster.iOS
         {
             //should make db filenames a string resource or something
             string documentsPath = Environment.GetFolderPath(Environment.SpecialFolder.Personal); // Documents folder
-
-
             string dbFileName = dbName;
 
             var dbPath = Path.Combine(documentsPath, dbFileName);
 
+            if (dbName == "ProductDB.db3")
+            {
+                if (!File.Exists(dbPath))
+                {
+                    File.Copy("ProductDB.db3", dbPath);
+                }
+            }
+            else if (dbName == "JobDB.db3")
+            {
+                if (!File.Exists(dbPath)) File.Create(dbPath);
+            }
+
             //to create a new db everytime use this - for debugging only 
-            File.Create(dbPath);
+            //File.Create(dbPath);
 
             return new SQLiteConnection(dbPath);
         }
